@@ -17,8 +17,8 @@
     }
   ],
   "require": {
-    "php": "^8.1",
-    "laravel/framework": "^8.0|^9.0|^10.0"
+    "php": "^8.3",
+    "laravel/framework": "^10.0|^11.0|^12.0"
   },
   "autoload": {
     "psr-4": {
